+++ resources/js/Admin/Components/index.js
@@ -22,7 +22,6 @@ export { FileUploader } from './FileUploader';
 export { ContentMediaFields } from './ContentMediaFields';
 export * from './CertificateSelector';
 export { PhoneInput } from './PhoneInput';
-export { YandexMapPicker } from './YandexMapPicker';
 export { CategoryTreeSelector } from './CategoryTreeSelector';
 export { default as RegionSelector } from './RegionSelector';
 export { default as SimpleWysiwyg } from './SimpleWysiwyg';